Relocating Sign Up Page
I've relocated the sign up page to the authentication folder as opposed to the splash_screen folder it was in previously. I've also created a basic form with fields to enter the email, password, first name, last name, phone number, address line 1 & 2, city, county and postcode. Finally a button to sign up is present.